Add 'res' query param to control ball mesh resolution
- Implemented 'res' query parameter parsing directly in BallMesh.
- Updated IcosahedronGeometry subdivision level to use the parsed 'res' value (defaults to 1).
- Included safety checks for valid integer input from URL.
